requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
pypdf
fastapi
uvicorn
streamlit
requests
rank-bm25
tiktoken